backend/requirements.txt
flask
flask-cors
python-dotenv
opencv-contrib-python-headless
pillow
numpy
ultralytics
insightface
onnxruntime
twelvelabs
